Create paged layout Fixes #9
